borrow and user management done

+++ src/main/java/com/zeynep/librarymanagementsystem/service/iml/UserServiceImpl.java
@@ -1,6 +1,9 @@
 package com.zeynep.librarymanagementsystem.service.iml;
 
 import com.zeynep.librarymanagementsystem.dto.UserDTO;
+import com.zeynep.librarymanagementsystem.exception.EmailAlreadyExistsException;
+import com.zeynep.librarymanagementsystem.exception.ISBNAlreadyExistsException;
+import com.zeynep.librarymanagementsystem.exception.UserNotFoundException;
 import com.zeynep.librarymanagementsystem.mapper.UserMapper;
 import com.zeynep.librarymanagementsystem.model.Book;
 import com.zeynep.librarymanagementsystem.model.User;
@@ -9,7 +12,7 @@ import com.zeynep.librarymanagementsystem.service.UserService;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
+
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -38,6 +41,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserDTO registerUser(UserDTO userDTO) {
+        if (userRepository.existsByEmail(userDTO.getEmail())) {
+            throw new EmailAlreadyExistsException("A User with the Email " + userDTO.getEmail() + " already exists.");
+        }
         // Convert DTO to Entity
         User user = userMapper.toEntity(userDTO);
         // Encode password before saving
@@ -51,7 +57,7 @@ public class UserServiceImpl implements UserService {
     @Override
     public UserDTO getUserById(Long id) {
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() -> new UserNotFoundException("User with ID " + id + " not found."));
         return userMapper.toDTO(user);
 
     }
@@ -65,19 +71,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserDTO updateUser(Long id, UserDTO updatedUserDTO) {
-        // Get the current logged-in user from the context
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String currentUserRole = authentication.getAuthorities().stream()
-                .map(grantedAuthority -> grantedAuthority.getAuthority())
-                .findFirst().orElse("");
-
-        // Check if the current user is a librarian
-        if (!currentUserRole.equals("ROLE_LIBRARIAN")) {
-            throw new RuntimeException("Only librarians can update user information.");
-        }
 
         return userRepository.findById(id).map(user -> {
             // Map DTO to Entity to update
+
             user.setName(updatedUserDTO.getName());
             user.setEmail(updatedUserDTO.getEmail());
             user.setRole(updatedUserDTO.getRole());
@@ -88,11 +85,14 @@ public class UserServiceImpl implements UserService {
             User updatedUser = userRepository.save(user);
             // Map updated entity back to DTO
             return userMapper.toDTO(updatedUser);
-        }).orElseThrow(() -> new RuntimeException("User not found"));
+        }).orElseThrow(() -> new UserNotFoundException("User with ID " + id + " not found."));
     }
 
     @Override
     public void deleteUser(Long id) {
+        if(!userRepository.existsById(id)) {
+            throw new UserNotFoundException("User with ID " + id + " not found.");
+        }
         userRepository.deleteById(id);
     }
 }
